feat: extract HTTP server to internal package and add runnable server
Refactor server implementation for production deployment:

- Move HTTP server implementation to internal/server package
- Create runnable server binary in cmd/server with CLI options
- Add backward compatibility wrapper in features/driver/http
- Maintain zero breaking changes - all tests pass unchanged

Key features:
- Production-ready server binary with configurable port
- Clean package structure following Go conventions
- Delegation pattern preserves existing public API
- Comprehensive documentation and usage examples

Server can now be deployed standalone while maintaining
in-process testing capabilities for development.
