+++ conversation-summaries/2025-09-17-runnable-server-extraction.md
@@ -1,0 +1,122 @@
+# Conversation Summary: Runnable Server Implementation
+
+**Date:** 2025-09-17
+**Session Duration:** ~10 minutes
+**Branch:** sirockin-add-http
+
+## Overview
+Extracted HTTP server implementation to an internal package and created a runnable server binary in cmd folder, while maintaining backward compatibility and ensuring all tests continue to pass.
+
+## Tasks Completed
+
+### 1. Package Structure Refactoring
+- **Extracted server implementation** to `internal/server/server.go`
+- **Updated package declaration** from `http` to `server`
+- **Maintained clean separation** between internal implementation and public API
+- **Preserved existing public interface** through delegation pattern
+
+### 2. Runnable Server Creation
+- **Created** `cmd/server/main.go` - production-ready server binary
+- **Added command-line options**: `-port` flag for configurable port
+- **Implemented proper logging**: Startup messages and endpoint documentation
+- **Added error handling**: Clean failure modes and informative messages
+
+### 3. Backward Compatibility Maintenance
+- **Updated** `features/driver/http/server.go` to delegate to internal package
+- **Preserved public API**: `NewServer()` function maintains same signature
+- **Zero breaking changes**: All existing tests pass without modification
+- **Maintained import paths**: Tests continue to work seamlessly
+
+### 4. Documentation Addition
+- **Created** `cmd/server/README.md` with usage instructions
+- **Documented API endpoints** with curl examples
+- **Explained architecture** and build/run procedures
+
+## Technical Implementation Details
+
+### Package Architecture
+```
+cmd/
+├── server/
+│   ├── main.go          # Runnable server binary
+│   └── README.md        # Usage documentation
+
+internal/
+└── server/
+    └── server.go        # HTTP server implementation
+
+features/driver/http/
+└── server.go            # Public API wrapper (delegates to internal)
+```
+
+### Delegation Pattern
+```go
+// Public API (backward compatible)
+func NewServer(app driver.ApplicationDriver) *server.Server {
+    return server.NewServer(app)
+}
+```
+
+### Server Features
+- **Configurable port**: Default 8080, customizable via `-port` flag
+- **Clean startup**: Informative logging with endpoint documentation
+- **Production ready**: Proper error handling and graceful failure
+- **Zero dependencies**: Uses only standard library and domain logic
+
+## Files Created/Modified
+- **NEW**: `internal/server/server.go` - Server implementation
+- **NEW**: `cmd/server/main.go` - Runnable server binary
+- **NEW**: `cmd/server/README.md` - Documentation
+- **MODIFIED**: `features/driver/http/server.go` - Delegation wrapper
+
+## Usage Options
+
+### Development/Testing
+```bash
+# In-process testing (existing)
+go test ./features
+
+# Both domain and HTTP tests pass identically
+```
+
+### Standalone Server
+```bash
+# Development mode
+go run ./cmd/server
+go run ./cmd/server -port=3000
+
+# Production build
+go build -o server ./cmd/server
+./server -port=8080
+```
+
+## Test Results
+- **Domain Tests**: ✅ 4 scenarios, 14 steps passing
+- **HTTP Tests**: ✅ 4 scenarios, 14 steps passing
+- **Build Status**: ✅ All packages compile cleanly
+- **Server Binary**: ✅ Successfully creates executable
+
+## Architecture Benefits
+
+### Separation of Concerns
+- **Internal package**: Implementation details isolated
+- **Public API**: Stable interface for testing
+- **Runnable binary**: Production deployment ready
+
+### Zero Breaking Changes
+- **Existing tests**: Continue to work unchanged
+- **Public API**: Preserved through delegation
+- **Import paths**: No modifications required
+
+### Deployment Options
+- **In-process**: Testing and development
+- **Standalone**: Production server deployment
+- **Docker ready**: Can be containerized easily
+
+## Next Steps
+The implementation provides:
+- ✅ Production-ready HTTP server binary
+- ✅ Clean package structure following Go conventions
+- ✅ Maintained test compatibility and coverage
+- ✅ Clear documentation for server usage
+- ✅ Foundation for containerization and deployment
